Modified xml2Src/nanohttp.c (and xml2Src/nanohttp.h) to parse, and to paginate, mjpeg streams.
coders/url.c also needs to paginate the stream. The open socket connection mustn't be closed & re-opened per frame, in this case, instead the handle is passed to the caller (Magic++/Image.cpp), which takes ownership. I wasn't able to identify a better approach to this.

Existing use of single-frame acquisition is not affected by these changes.

I've tested this on Windows7, Ubuntu 18.04 and Scientific Linux 7.3.

I've tested it with 2 cameras I have here, and a selection of publicly accessible Internet cameras.